// src/app/shared/types/article.interface.ts
import { ProfileInterface } from "./profile.interface";

export interface ArticleInterface {
    title: string;
    slug: string;
    body: string;
    createdAt: string;
    updatedAt: string;
    tagList: string[];
    description: string;
    favorited: boolean;
    favoritesCount: number;
    author: ProfileInterface
}